11_7_25/web_scrapping_agent/requirements.txt
langchain
langchain-community
beautifulsoup4
requests
python-dotenv
